server/requirements.txt
flask
python-dotenv
flask-cors
pymongo
bcrypt
requests